+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Note that since we don't clearly distinguish between a public and private interf
 
 ## [Unreleased]
 
+- Add parameter for to display aromatic bonds as dashes next to solid cylinder/line.
+
 
 ## [v2.0.7] - 2021-06-23